+++ tests/dtale/test_charts.py
@@ -30,3 +30,11 @@ def test_build_agg_data():
 def test_weekday_tick_handler(unittest):
     unittest.assertEqual(chart_utils.weekday_tick_handler([1, 2, 3], 'date|WD'), ['Tues', 'Wed', 'Thur'])
     unittest.assertEqual(chart_utils.weekday_tick_handler([1, 2, 3], 'foo'), [1, 2, 3])
+
+
+@pytest.mark.unit
+def test_valid_chart():
+    assert not chart_utils.valid_chart(chart_type='line', x='a', y='b', agg='rolling')
+    assert not chart_utils.valid_chart(chart_type='line', x='a', y='b', agg='rolling', window=10)
+    assert not chart_utils.valid_chart(chart_type='line', x='a', y='b', agg='rolling', rolling_comp='sum')
+    assert chart_utils.valid_chart(chart_type='line', x='a', y='b', agg='rolling', window=10, rolling_comp='sum')